% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\docType{data}
\name{KafkaConsumer}
\alias{KafkaConsumer}
\title{Creates KafkaConsumer object (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaConsumer
}
\description{
Creates KafkaConsumer object (R6 class)
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
consumer$poll(interval = 100)
}
}

\section{Methods}{

\describe{
  \item{\code{consumer = KafkaConsumer$new(topic_name, bootstrap_servers, group_id, config = list())}}{}
  \item{\code{consumer$poll(interval = 100)}}{Fetch data for the topics or partitions specified
  using one of the subscribe/assign APIs. On each poll, consumer will try to use the last consumed offset as the starting offset and fetch sequentially.
  The last consumed offset can be manually set through #seek(TopicPartition, long) or automatically
  set as the last committed offset for the subscribed list of partitions}
}
}

\section{Arguments}{

\describe{
  \item{consumer}{A \code{KafkaConsumer} object}
  \item{topic_name}{name of a Kafka topic to read messages from}
  \item{broker_list}{bootstrap_servers for a Kafka cluster}
  \item{config}{ConsumerConfig parameters (in \code{key = value} format) for kafka client -
  see them here \url{https://kafka.apache.org/0100/javadoc/index.html?org/apache/kafka/clients/consumer/ConsumerConfig.html}}
  \item{group_id}{group_id for a Kafka client - will be used to understand offsets}
  \item{interval}{\bold{integer >= 0} scalar value. The time, in milliseconds, spent waiting in poll if data is not available in the buffer.
  If 0, returns immediately with any records that are available currently in the buffer, else returns empty.}
}
}
\keyword{datasets}

