% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.R
\docType{data}
\name{KafkaProducer}
\alias{KafkaProducer}
\title{Creates KafkaProducer object (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaProducer
}
\description{
Creates KafkaProducer object (R6 class)
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
producer$send(messages)
}
}

\section{Methods}{

\describe{
  \item{\code{producer = KafkaProducer$new(broker_list, client_id = "RscalaKafkaClient", config = list())}}{Creares
  instance of KafkaProducer}
  \item{\code{producer$send(messages)}}{Asynchronously send a record to a topic.}
}
}

\section{Arguments}{

\describe{
  \item{producer}{A \code{KafkaProducer} object}
  \item{client_id}{string, identifier of the client}
  \item{topic}{name of a Kafka topic to send messages}
  \item{broker_list}{broker list for a Kafka cluster}
  \item{config}{ProducerConfig parameters (in \code{key = value} format) for kafka client -
  see them here \url{https://kafka.apache.org/0100/javadoc/index.html?org/apache/kafka/clients/producer/ProducerConfig.html}}
  \item{messages}{\bold{character} strings - messages to send to kafka}
}
}
\keyword{datasets}

